refactor: consolidate auth endpoints into frappe/core/api/auth.py
Move the auth-related whitelisted endpoints - logout/web_logout
(handler.py, keeping plain import aliases there for the bare-cmd
/api/method/logout shorthand), login_via_token/send_login_link/
login_via_key (www/login.py), get_logged_user (auth.py),
reset_otp_secret (twofactor.py), sessions.clear (renamed clear_cache)
and the User module's password/signup/impersonation endpoints
(update_password, test_password_strength, verify_password, sign_up,
reset_password, change_password, generate_keys -> generate_api_keys,
impersonate, clear_session) - into frappe/core/api/auth.py, marked
@frappe.public(group="Auth") with full annotations and docstrings.

Old dotted paths (including the login/signup/update-password pages'
cmd strings until this commit's JS/template updates) keep working via
aliases; test patch targets follow the functions to the new module.
