+++ frappe/public/js/frappe/ui/toolbar/toolbar.js
@@ -254,7 +254,7 @@ $.extend(frappe.ui.toolbar, {
 
 frappe.ui.toolbar.clear_cache = frappe.utils.throttle(function () {
 	frappe.assets.clear_local_storage();
-	frappe.xcall("frappe.sessions.clear").then((message) => {
+	frappe.xcall("frappe.core.api.auth.clear_cache").then((message) => {
 		frappe.show_alert({
 			message: message,
 			indicator: "info",
