+++ frappe/tests/test_api.py
@@ -549,7 +549,7 @@ class TestAPIResponse(FrappeAPITestCase):
 
 
 def generate_admin_keys():
-	from frappe.core.doctype.user.user import generate_keys
+	from frappe.core.api.auth import generate_api_keys as generate_keys
 
 	generate_keys("Administrator")
 	frappe.db.commit()
