+++ frappe/www/update-password.html
@@ -169,7 +169,7 @@ frappe.ready(function() {
 
 		frappe.call({
 			type: "POST",
-			method: "frappe.core.doctype.user.user.update_password",
+			method: "frappe.core.api.auth.update_password",
 			btn: update_button,
 			args: args,
 			statusCode: {
